moved files and made media queries
