+++ code/src/components/Form.css
@@ -1,7 +1,8 @@
-.question-header {
+/*.question-header {
     font-size: 16px;
     padding: 10px;
 }
+
 .container {
-display: block;
-}
+    display: block;
+}*/
